rename and re-structre to allow proper install
