+++ src/lidar_tools/__init__.py
@@ -1,6 +1,6 @@
 from lidar_tools import filter_percentile
 from importlib.metadata import version as _version
 
-__version__ = _version("grid_pc")
+__version__ = _version("lidar_tools")
 
 __all__ = ["filter_percentile"]
